+++ neuropsydia/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.0"
+__version__ = "1.0.3"
 
 from .path import *
 from .core import *
@@ -11,4 +11,5 @@ from .ask import *
 from .choice import *
 from .meta import *
 from .io import *
+from .procedures import *